sending errors and displaying them on screen actually working
