+++ frontend/src/components/Root.jsx
@@ -1,5 +1,7 @@
+//router
 import { Outlet } from 'react-router-dom';
 
+// components
 import NavBar from './Navbar/Navbar';
 
 const Root = () => {
